requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
firecrawl-py==0.0.16
openai==1.99.9
